service/keycloak-sample-app/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.3"}
pydantic-settings = "^2.3.4"
requests = "^2.32.3"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
fastapi-sessions = "^0.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
